Adjust material utils
The names for reference and retopo materials are now stored in module level constants.
These constants are expected by get_material.
It now checks whether the material is valid, and if it isn't, resets it and does the setup.
Also I moved the TODO for node position, and renamed assign_material to set_material.