Get wire settings on modifier creation
This does mean it redundantly sets the settings in their update methods the first time, but afterwards it won't overwrite any settings inside this function because it just returns the existing modifier, so eh it's fine.
Also I'm taking the modifier creation out of the else statement because it's silly, and changing the default value for even offset to match the blend file I got.
And adding a default for offset because it was missing (0.0 doesn't match the 1.0 I was given, but I think this looks better).