Replace retopo intensity with reference opacity
You know how you change retopo intensity? Adjust the color by dragging the slider to the right of the color wheel.
For reference, I could perhaps read the alpha of the color and use that as the alpha of the principled node, but I don't think that's very intuitive; the user wouldn't know that the functionality exists if we don't tell them; so having an explicit slider makes sense here.
Also I'm adding an extra separator between reference and retopo material settings.
I might give them their own subpanels later.