Overhaul settings panel
This panel was getting a little long, so I'm using a column now to squeeze things down a bit.
I've grouped together reference and retopo color, but not wire color.
Wire color instead lives with the rest of the wire settings, and vertex size, because they're all related.
I also added world strength to the bottom of the panel.