Make it so get_material/modifier only create when told to
It might be a bit annoying for the addon to create materials if you so much as touch a related setting, so now the settings don't create materials or modifiers anymore, but just do nothing when the material/modifier they're looking for isn't there.
The add material operators explicitly tell the utils to create materials (if they don't exist already).