Make wire thickness property nicer to interact with
I took inspected the settings of the property on the wireframe modifier some more, and found these are what make it feel good.
Did you know that step can be a float? I thought it had to be an int.
Works fine in 2.93 and 2.83, so I dunno why the documentation doesn't mention this.
The documentation says it should be an int, which is then divided by 100. In practice it is not divided by 100; 1 just means 1.