Add displace and solidify props to settings panel
I think this approach of putting the checkbox on the row with the slider is nice; though it wouldn't work if there was more than one value per modifier.
Therefore I'm moving wireframe color up with the other material colors.
I'm also getting rid of the words "Color" and "Thickness" from the UI, because I think it speaks for itself, and was getting too wide.
Because I want to be able to support Blender 2.83 LTS (it's still pretty popular I think), I implemented a method for creating the split row approach that will work on both 2.9 and 2.8.
Maybe I should use a version check instead of a try block, as it would probably be more performant. I'll worry about it later.