Adjust setup modifier utils
To match solidify, displace has to use a mid level of 0.
Also we want to use show_on_cage for displace and solidify.
I explicitly set it to False for wireframe, even though that modifier doesn't support show_on_cage anyway, just in case they add it in the future (as we will probably want it disabled for this modifier).