Replace warning with tooltip for add reference material
The warning felt a little overkill, but mostly it bothered me that you'd see this same draw panel in the bottom left after using the operator.
I think using a tooltip is good enough, and doesn't have this issue with the panel.
I'm also removing our retopo modifiers (if present) when adding the reference material.