Check for retopo modifiers in move modifiers operator
I think it's better for this operator to only be available if there is at least one of our retopo modifiers present on the object.
I'm not checking whether the object is valid first, because get_modifier already does that anyway.
This approach does mean it's doing the check three times, but I doubt that hurts performance at all.