Change retopo material from Emission to Principled BSDF
Emission is fine when the retopo mesh is inside the reference mesh, and sort of inherits its shading.
When using displace or solidify though, you'll see the retopo mesh as-is, which is much too flat.
So I'm changing it to a principled shader with settings which I think look nice.
Also I'm adding type hinting to color.