Support alpha on all materials
Also I decided to change the defaults, so reference is now opaque while retopo is translucent, I think this works better with offset.
I set blend mode based on alpha value, because opaque materials are visible behind translucent ones.
For this same reason, I always set the wireframe material to use alpha blend, because we don't want to see the wireframe through a translucent reference mesh.
I changed the node setup for back facing in the retopo material because alpha on the principled node is used for opacity now.
The wireframe material uses a principled node now as well (because it has alpha input, which is simpler than having to use a mix node and a transparent node).
I make it black by setting the base color to black and specular to 0, then put the color in emission.
As long as we don't need emission strength, this works in in 2.83.