Don't setup materials from draw context
It was trying to reset invalid materials from property getters, which were being called by draw code; that's not allowed.
Now I setup materials only when create is used, and otherwise if the material is invalid, just don't return it.
Also I added utilities for check and setup material because they're convenient to have.