Fix materials when possible
I made a utility to check whether the current context has write access to ID properties, which is necessary for fixing materials.
If the material is not valid, but the context allows for it, we fix it (this is for example from property setters).
If the material is not valid, but the context does not allow for it, we just don't return it, as there's no point trying to access and invalid material.