Add Mirror Modifier operator
This adds a mirror modifier at the top of the stack, or adjusts the first mirror modifier if there is one.
Took me a bit to figure out how to draw a BoolVectoryProperty the way it is here.
Also row_with_heading is back baby.
There's no sliders in the UI, so a standard redo approach performs fine.
I'm thinking though.. maybe I should just switch to the modifiers tab and let the user do it there?